Implemented logging.
In this section we create two new projects,  Contracts and LoggerService.
Contracts hold all the application level interfaces and individual services (such as Logger Service for now) will implement the interface(s). We then register the logger service with the main application via Dependency injection.
We also test the logger service by changing the Values Controller - auto generated code which is a throw away when we complete the solution.